Refactored flappy-judoka integration with the Updater
Implemented some optimisations and a code refactor to how the game is launched.

- using key constants for [J] key
- moved game launching to install.py view
- optimised killing with pkill
- code comments
