+++ kano_updater/ui/install_window.py
@@ -13,10 +13,9 @@ from threading import Thread, Lock
 
 from kano.gtk3.apply_styles import apply_styling_to_screen
 from kano.gtk3.kano_dialog import KanoDialog
-from kano.utils import is_model_2_b
 
 from kano_updater.utils import kill_apps
-from kano_updater.ui.paths import CSS_PATH, FLAPPY_PATH
+from kano_updater.ui.paths import CSS_PATH
 from kano_updater.commands.install import install
 from kano_updater.ui.progress import GtkProgress
 from kano_updater.ui.views.install import Install
@@ -33,6 +32,8 @@ class InstallWindow(Gtk.Window):
 
         Gtk.Window.__init__(self)
         # self.fullscreen()
+        # Gtk hack: set the width, height of the window to larger than screen
+        # resolution in order fix set_keep_below(True) which doesn't work with fullscreen
         screen = Gdk.Screen.get_default()
         width = screen.get_width()
         height = screen.get_height()
@@ -48,10 +49,8 @@ class InstallWindow(Gtk.Window):
         kill_apps()
 
         self.show_all()
+        self._install_screen.hide_game_play_label()
         self._set_wait_cursor()
-
-        self._install_screen._pgl.hide()
-        self._is_game_first_launch = True
 
         # For passing user input to the install thread
         self.user_input = None
@@ -61,29 +60,6 @@ class InstallWindow(Gtk.Window):
         self.user_input_lock.acquire()
 
         self._start_install()
-
-    def _launch_game(self, window=None, event=None):
-        try:
-            if event and event.get_keycode()[1] == 44:  # [J] key
-                if self._is_game_first_launch:
-                    self._is_game_first_launch = False
-                    self._first_game_launch()
-                else:
-                    self._relaunch_game()
-        except:
-            pass
-
-    def _first_game_launch(self):
-        self.set_keep_below(True)
-        os.system('{} &'.format(FLAPPY_PATH))
-        os.system('kill -9 `pgrep -f kano-feedback`')      # kill kano-feedback-widget
-        os.system('kill -9 `pgrep -f kano-world-widget`')  # kill kano-world share widget
-        os.system('lxpanelctl exit')                       # kill lxpanel
-        # minimise all windows, and restore focus to the Updater and then Flappy
-        os.system('wmctrl -k on && wmctrl -a "Updater" && wmctrl -a "Flappy"')
-
-    def _relaunch_game(self):
-        os.system('{} &'.format(FLAPPY_PATH))
 
     def _start_install(self):
         progress = GtkProgress(self)
@@ -151,14 +127,7 @@ class InstallWindow(Gtk.Window):
         Gtk.main_quit()
 
     def update_progress(self, percent, msg, phase_name, sub_msg=''):
-        self._install_screen.update_progress(percent, msg, sub_msg)
-
-        # enable flappy-judoka only for the RPI2
-        # enabling flappy-judoka launch only after these phases (when a reboot is iminent)
-        if is_model_2_b():
-            if phase_name in ['downloading', 'downloading-pip-pkgs', 'init', 'installing-urgent']:
-                self.get_toplevel().connect('key-release-event', self._launch_game)
-                self._install_screen._pgl.show()
+        self._install_screen.update_progress(percent, phase_name, msg, sub_msg)
 
         # FIXME Progress to next with the done
         if percent == 100:
